+++ app/src/main/java/com/example/course_project/data/db/cart/LocalCartDataSource.java
@@ -17,7 +17,7 @@ public class LocalCartDataSource implements CartDataSource {
     }
 
     @Override
-    public Single<Long> sumPriceInCart(long userId) {
+    public Single<Double> sumPriceInCart(long userId) {
         return cartDAO.sumPriceInCart(userId);
     }
